+++ src/Domain/Model/Happening.php
@@ -71,6 +71,9 @@ class Happening implements TimeRangeInterface
     /** @var bool */
     private $interactiveWebinar;
 
+    /** @var bool */
+    private $videoWebinar;
+
     /** @var null|string */
     private $webinarSessionId;
 
@@ -88,6 +91,7 @@ class Happening implements TimeRangeInterface
         ?string $invitationCode = null,
         bool $webinar = false,
         bool $interactiveWebinar = false,
+        bool $videoWebinar = false,
         ?string $liveUrl = null
     ) {
         $this->event = $event;
@@ -105,6 +109,7 @@ class Happening implements TimeRangeInterface
         $this->products = new ArrayCollection();
         $this->webinar = $webinar;
         $this->interactiveWebinar = $interactiveWebinar;
+        $this->videoWebinar = $videoWebinar;
         $this->liveUrl = $liveUrl;
     }
 
@@ -233,6 +238,7 @@ class Happening implements TimeRangeInterface
         ?int $limitParticipant,
         bool $webinar,
         bool $interactiveWebinar,
+        bool $videoWebinar,
         ?string $invitationCode = null,
         ?string $liveUrl = null
     ) {
@@ -245,6 +251,7 @@ class Happening implements TimeRangeInterface
         $this->invitationCode = $invitationCode;
         $this->webinar = $webinar;
         $this->interactiveWebinar = $interactiveWebinar;
+        $this->videoWebinar = $videoWebinar;
         $this->liveUrl = $liveUrl;
     }
 
@@ -460,12 +467,17 @@ class Happening implements TimeRangeInterface
 
     public function isWebinar(): bool
     {
-        return $this->webinar || $this->interactiveWebinar;
+        return $this->webinar || $this->interactiveWebinar || $this->videoWebinar;
     }
 
     public function isInteractiveWebinar(): bool
     {
         return $this->interactiveWebinar;
+    }
+
+    public function isVideoWebinar(): bool
+    {
+        return $this->videoWebinar;
     }
 
     public function getLiveUrl(): ?string